refactor: components/@shared 파일 import 하나로 병합
