+++ src/components/@shared/index.ts
@@ -1,0 +1,7 @@
+export { default as Button } from './Button';
+export { default as CheckBox } from './CheckBox';
+export { default as Input } from './Input';
+export { default as Link } from './Link';
+export { default as Loading } from './Loading';
+export { default as NumberInput } from './NumberInput';
+export { default as ShoppingCart } from './ShoppingCart';
